requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
freezegun==1.4.0
gunicorn==21.2.0
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
